pi/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
paho-mqtt==1.6.1
python-dotenv==1.0.1
requests==2.32.3
python-multipart==0.0.12
